app/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
black==23.7.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.6
crashtest==0.4.1
cryptography==41.0.2
distlib==0.3.7
dnspython==2.4.0
dulwich==0.21.5
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.100.0
fastapi-users==12.1.0
fastapi-users-db-sqlalchemy==6.0.0
filelock==3.12.2
flake8==6.0.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.3.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
keyring==23.13.1
lockfile==0.12.2
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
more-itertools==10.0.0
msgpack==1.0.5
multidict==6.0.4
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.9.1
pluggy==1.2.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
ptyprocess==0.7.0
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.0.3
pydantic-settings==2.0.2
pydantic_core==2.3.0
pyflakes==3.0.1
PyJWT==2.7.0
pyproject_hooks==1.0.0
pytest==7.4.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-telegram-bot==20.4
rapidfuzz==2.15.1
referencing==0.30.0
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.9.2
rsa==4.9
ruff==0.0.280
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tomlkit==0.11.8
trove-classifiers==2023.7.6
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.1
virtualenv==20.24.2
webencodings==0.5.1
xattr==0.10.1
yarl==1.9.2
zipp==3.16.2
